make it less based on nix

+++ run.sh
@@ -4,14 +4,25 @@ set -euo pipefail
 export LANG=C
 ROOT="$( cd $(dirname "$0") && pwd)"
 
+guardForBins() {
+    for cmd in "$@"; do
+        if ! command -v "$cmd" &> /dev/null; then
+            >&2 echo "$cmd is required"
+            return 1
+        fi
+    done
+}
+
 getOutputFolder() {
     local OUT
     OUT="${ROOT}/out/$(hostname)"
     if [[ -f /sys/firmware/acpi/platform_profile ]]; then
         OUT="$OUT-$(cat /sys/firmware/acpi/platform_profile)"
     fi
-    if acpi -a | grep -v "on-line" -q; then
-        OUT="$OUT/onBattery"
+    if guardForBins "acpi"; then
+        if acpi -a | grep -v "on-line" -q; then
+            OUT="$OUT/onBattery"
+        fi
     fi
     OUT="$OUT/$(date +%Y-%m-%d)"
     mkdir -p "$OUT"
@@ -28,7 +39,7 @@ writeStats() {
 
 benchFio() {
     local OUT="$1/fio"
-    if [[ ! -d "$OUT" ]]; then
+    if [[ ! -d "$OUT" ]] && guardForBins "fio"; then
         mkdir -p "$OUT"
         fio --name TEST --eta-newline=5s --filename=temp.file --rw=randread --size=2g --io_size=10g --blocksize=4k --ioengine=libaio --fsync=1 --iodepth=1 --direct=1 --numjobs=32 --runtime=60 --group_reporting |
             tee -a "$OUT/random-4K-reads"
@@ -40,7 +51,7 @@ benchFio() {
 
 benchGlmark() {
     local OUT="$1/glmark"
-    if [[ ! -f "$OUT" ]]; then
+    if [[ ! -f "$OUT" ]] && guardForBins "glmark2"; then
         glmark2 |
             tee "$OUT"
         git add "$OUT"
@@ -48,11 +59,15 @@ benchGlmark() {
 }
 
 benchFromFile() {
+    guardForBins "perf" || return 0
     local OUT="$1"
     local script="$2"
     local bn="$(basename "$script")"
     local bOUT="$OUT/${bn%.*}"
+
+    guardForBins $("$script" requirements "$bOUT") || return 0
     mkdir -p "$bOUT"
+
     {
         "$script" prepare "$bOUT"
         "$script" run "$bOUT"
@@ -74,6 +89,12 @@ benchesFromFiles() {
         done
 }
 
+benchesFromFunctions() {
+    local OUT="$1"
+    benchFio "$OUT"
+    benchGlmark "$OUT"
+}
+
 main() {
     local OUT
     OUT="$(getOutputFolder)"
@@ -81,11 +102,9 @@ main() {
 
     if [[ $# -eq 0 ]]; then
         benchesFromFiles "$OUT"
-        benchFio "$OUT"
-        benchGlmark "$OUT"
+        benchesFromFunctions "$OUT"
     elif [[ "$1" == "--other" ]]; then
-        benchFio "$OUT"
-        benchGlmark "$OUT"
+        benchesFromFunctions "$OUT"
     else
         for script in "$@"; do
             benchFromFile "$OUT" "$(readlink -f "$script")"
